Send reply messages by editing received message file
To respond to message number N, edit N.txt. The text in N.txt
will appear quoted. (Provided N is indeed a *received* message,
not an outgoing one.) Any unquoted text added will be sent as
a reply message when the file N.txt is closed. Edits to the
file will be lost, but will be present in a new message file
M.txt corresponding to the reply message M.