Add “out” file to all chat directories
All new messages are appended to the new “out” file, prefixed by their author.

I've chosen to never return an empty result for read syscalls. That
means

	cat out

and

	tail -f out

have the same behaviour, and avoids having to poll the file for new
data. That works well under Plan 9 but causes temporary problems
under Linux; but I can live with it. The problem is that the cat
process can't be immediately killed under Linux. As long as no new
messages arrive, the process will be in uninterruptible sleep state.
So, killing the process just enqueues a signal to be delivered to the
process once it gets unstuck (once it finishes the read syscall). A
killed process won't exist until a new message arrives!